Remove short circuit in `app.$process()` if no routes found. (#396)
fixes #393 

The returned `{ status: 200 }` from `app.$process` is not handled by the
plugin, and so the http response is not set causing the server to
"hang". By not short circuiting, the program execution continues leading
to `onActivityResponse` event being called and the HTTP response being
set to default status of 200.

- [x] **Is it an issue in .NET SDK?** Nope, there is no short-circuiting
being done like this.
- [x] **Is it an issue in Python SDK?** Yes, it should be fixed:
https://github.com/microsoft/teams.py/pull/218